Cleanup and type change for temps

+++ main.cpp
@@ -112,10 +112,9 @@ vector<float> orderTemp(const vector<float> temps){
     vector<float> order;
     order.reserve(temps.size());
 
-    //maybe error was here in loop, error with iteration was present here
     for (int i = 0; i <= 20; i++){
         //make our i into tenth form
-        double temp = 97.0 + (i * .1);
+        float temp = 97.0 + (i * .1);
 
         //iterate and find all occurences of a specific temp
         auto it = occurences.find(temp);
@@ -154,7 +153,6 @@ int longSeq(const vector<int>& vals){
     return longest_seq;
 }
 
-//Maybe add print statements inside functions...
 
 int main(){
 //How Im declaring the Player arrays as type/example given is most likely written in Ruby
